requirements.txt
requests
beautifulsoup4
lxml
PyPDF2
faiss-cpu
google-generativeai
openpyxl
pandas
python-dotenv